Use value composition.
